Simplified approach for updating abilities
The new approach makes all of the updates in one pass, avoids reduce, and makes the overall flow of the code clearer. Additionally, we shouldn't filter out abilities that haven't changed from the `bulkPut`. Even though we can try to skip bulkPuting abilities that had no changes, we gain a lot vs the added complexity.