Clarify control compatibliity for external devices and names
Expand the usage of hardware wallets to instead talk about external
devices. WalletConnect as an external device is a potential future use
case, and is a good illustration of a situation where the signer is
really never known to be transaction-compatible on any network other
than the one it was added for.

Additionally, add a note about name ownership to indicate it's not
currently tracked, some stipulations if we do want to use it, and a flag
that it's probably not worth the effort right now.