Improve wording in README
Improved wording, fixed typos. Also removed the line about running all tests (we
can't run all e2e tests with the same command expecting all of them to
pass - the environments on which fork-based tests are designed for differs from
the environment where rest of the tests iss supposed to be run.