Catch and log errors in Ledger onConnection
Previously if there was an error with a connected Ledger during
connection setup, we bubble it out of the Ledger service. Now, that
error is caught, logged, and the Ledger is treated as having just
disconnected. This gives consumers a more accurate picture of the state
we are in (namely, that we can't properly connect to the Ledger).

In the future we could also emit some sort of connection error and tell
the user we had an active error connecting instead of a passive
inability to detect the Ledger, but not going down that rabbit hole
right now.