Create separate build artifact for testing on Mainnet fork
In order for the Taho extension to work on the forked Mainnet network it needs
to be built with the `USE_MAINNET_FORK` environment variable set to `true`. As
we're running the E2E on fork in the CI, we need to create a dedicated build
package in the CI (built with that variable configured) that will be used to set
up the testing environment.