Modify handling of the expected balance after tx
Previously after tx was made we've been checking if the balance gets updated to
a value smaller (or equal) than the provided `expectedMaxBalance` (which on fork
we were setting to balance minus the transaction amount). It makes more sense to
change how we check the balance and verify if it matches provided RegExp value.
This way for fork we can match with exact value (the balance should be the same
on each run) and for tests run on 'real' networks we can use more general
RegExps, as there we can't predict the resulting balance accurately.