Change data type of some arguments to RegExp
In the couple of helper functions that we use in E2E tests we are expecting
string parameters in particular format which we later transform to a RegExp
syntax. We do that in situations where we need to build over the provided
argument to construct more complicated RegExp we're matching aganst or when we
use `getByText` locator. Previously we've been doing that also in some places
where we didn't need to. Now we're switching to using RegExp arguments where
possible.