Don't stor Playwright artifacts in job details
So far, when running E2E tests in CI, we were storing some Playwright artifacts
in job details if some of the tests have failed. The artifacts include
screenshots from the tested pages, made during each action, so if we import
account from seed in the tests, the seed will be captured as well. As in the
recent commits we've been adding tests of transactions which rely on import of
an existing account, we're now risking leeking the seed phrase of that account.
Eventually we want to switch to importing accounts using JSON, which will not
expose the sensitive data. Before this kind of account import becomes supported
in Taho, we're avoiding leeking the phrase by temporarly turning off the storing
of Playwright artifacts.