Move top and tab bar management into CorePage
This was being handled in Popup, but these bars are really a property of
the page itself—for example, the transaction signing page has a
different top bar and no tab bar, pages like Settings have the tab bar
but no top bar, and the wallet has both.

While this commit does not move all bar handling into the pages
themselves, it does start moving the management of them into the page
definitions by having `CorePage` take charge of rendering the tab and
top bars (or not) depending on the parameters it is passed.