Allow SharedToggleButton to have a different direction and colors
`onColor` and `offColor` parameters allow modifying the colors that the
toggle's bulb takes on when it's turned on as well as when it's turned
off. `leftToRight`, defaulting to `true`, indicates that the toggle is
off on the left and on when on the right. Passing it as `false` flips
the two horizontally.