Add DAppConnection component to top of wallet page
This component allows the user to manage the active dApp connection, as
well as allowing them to always toggle the default wallet to/from Taho.

The popup window with information on connecting to dApps when they are
not connected yet needs some refinement for the new default connection
approach, but that will be done separately.