Immediately hide exiting pages during transitions
In practice, the only page transition that is visible, due to
positioning/layout specifics, is the incoming page. Outgoing pages are
pushed under the incoming page during layout, so their transitions
aren't visible in the first place. Animations were retargeted to only
the `main` element, however, which meant that there *was* a briefly
visible version of the outgoing page visible overlapping the bottom of
the incoming page.

This commit drops all transitioning for the outgoing page, instead
setting it to immediately transparent for the duration of the transition
(after which the page is completely removed from the DOM).