Add DAppConnectionInfoBar to the top of the Connect dApp page
This component is the top bar when connecting to a dApp, and allows
switching the default wallet to something other than Taho while
connecting, instead of requiring opening the popover separately. Similar
functionality was previously rendered as a button in the middle of the
connection page (and still is when `ENABLE_UPDATED_DAPP_CONNECTIONS` is
off).