Allow default toggle to be used as a static button
The new `alwaysForceSelection` prop can be set to `taho` or `other` to
indicate that the toggle should always be rendered in a set state, and
clicking it should force that default-wallet state in the background.
This is used when instead of a toggle we want to show the user a
specific way to set the value, for example when informing them that to
use MetaMask they should set the toggle to MetaMask.