Always route new requests to the current provider
The way that the provider proxy was set up, changing the current
provider would not route new requests to that provider, instead
requiring the dApp to reconnect. To allow transparent forwarding to the
new default when the connection switches, the proxy now always forwards
to the current provider (when available).