Close dApp connection page when default wallet changes
Only when updated dApp connections are enabled, the dApp connection
popup auto-closes (after a 300ms delay to let the user see the toggle
animate). The expectation is that the new default wallet will pop up
immediately as the window provider replays the connection request from
Taho to the new default.