Add dismissable items to preferences
These are items that are displayed to the user either exactly once, or
once until the user closes them. The default connection popover is
supported, but not yet hooked up.

This is designed to be a useful primitive for any new-feature banner, or
first-time-use banner. Like many items in preferences, the current list
of already-shown dismissable items is initialized when the extension
starts, and then updated along the way based on UI actions.