Create E2E tests for verified/unverified tokens
This commit introduces E2E test that checks the functiolality of
verified/unverified tokens. Following general steps are executed:
1. Import account
2. Enable `Show unverified assets`
3. Hide asset
4. Trust asset
5. Hide trusted asset
A number of checks is performed during each step.

The commit also introduces helper functions and adds `data-testid` attribute to
a couple of DOM elements.

As some of the code is similar or identical as in the
https://github.com/tahowallet/extension/pull/3418 PR which is yet not merged to
`main`, some conflicts may arise and will need to be resolved before this change
lands on `main`.

Also some changes will need to be made once
https://github.com/tahowallet/extension/pull/3195 gets merged to `main`, as this
PR solves a bug causing failures in the tests (the failing part of the test was
temporarily commented out).