Render top bar default toggle in same direction as other toggles
The default toggle was being rendered with "Taho as default" being on
the left. After some discussion, the fact that it was reversed compared
to other toggles (where "on" is on the right) isn't ideal. This commit
puts "Taho as default" on the right, and drops the `leftToRight`
parameter added to `SharedToggleButton` to make that possible in the
first place.