Configure Hardhat project for running extension on a forked Mainnet
In this commit we create a Hardhat project that allows for running Taho
extension on a forked Mainnet. As part of this we're configuring a `hardhat`
network in the Hardhat config file. The fork will begin on a block specified in
the `blockNumber` parameter. The `url` parameter should hold an url to a
JSON-RPC node that's based on your personal Infura/Alchemy API key (e.g.
`https://mainnet.infura.io/v3/<key>`). The suggested value for `chainId` is
`1337`.
The steps to run the network are described in the README file.
Any actions performed on the `hardhat` network will only be applied locally,
they will not affect the real Mainnet network.