Execute E2E tests more frequently
So far, we've been running E2E tests only when a PR was changing feature flag,
contained `e2e` in the name of the feature branch or got merged to `main`. This
prevented us from noticing changes braking E2E tests fast enough - we sometimes
were noticing such changes only after they got already merged to `main`.
In this commit we modify CI config in a way that most of the E2E tests will be
run on every PR update. We also allow for specifying tests which shouldn't be
run this often - such tests should be marked with `@slow` tag in the test name
(read more here: https://playwright.dev/docs/test-annotations#tag-tests). We
may want to use this feature when we'll have tests that require spending some
funds.