Only auto-dismiss Connect popup when default transitions to false
Previously we were auto-dismissing if isDefaultWallet was ever false.
This meant if the user toggled the wallet to being not-default, then
tried to connect to a dApp, it would immediately dismiss—for example
when Taho-as-default is disabled and the user tries to connect to Taho
as Injected or as Taho on a website.

Now, the initial value when the popup was rendered is also checked, so
that the auto-dismissal only occurs if the popup was rendered when the
Taho-as-default was true, and then switched to false. This makes the
auto-dismissal work when the user changes the default setting in the
popup.