Tag costly test with the `@expensive` tag
The not-fork-based test validating transactions spends some Goerli, funds. As we
don't own an unlimited amount of Goerli ETH, we don't want to run this test on
every push to the PR feature branch. To limit how often it's run, we tag it with
the `@expensive` tag. Tests with this tag will be executed in the CI only in
limited situations.