Track vault migrations and errors generically
Vault migration is no longer tracked as Argon2 specifically, but
generically for all migrations. Already-migrated vaults are not tracked,
and the migration function return value reflects that no migration was
performed. Additionally, error messages are bubbled out of the migration
function and reported up to the caller.

The main outcome here is that PostHog migration events include the
migrated-to version, and PostHog migration failure events include the
error message. This will leave us open to future migrations, and will
let us know if there are certain failures that are happening broadly
that we may be able to do something about.

Notably, wrong passwords will be tracked as migration errors if a wrong
password is typed with an older vault version in the mix. Mitigating
this may or may not be a good idea.