Don't run e2e tests when workflow is triggered by push of tag
In commit 780d6d4eb817d05300a8458e534a10d1fa12e4ca we've turned off uploading of
build artifact to workflow details in case when the workflow is triggered by a
push of tag. Later in 168d945c0e138c250ca4edac27f9071a59fb0d83 we've decided to
change conditions of e2e tests execution to run the tests more frequently. But
we've forgotten that we can't run the tests when workflow is triggered by a tag
push, as there is no build artifact available. Now we're updating the triggers
of e2e tests to take that under consideration.