Refactor assets slice to use object instead of array
As Redux was having serious performance problems when we were using
an array to store all assets let's refactor that slice so it is now storing
assets as an object that is indexed by assets' symbols.
Seems like this is working great with redux and we are able to reduce
CPU usage as we were updating assets very often.