Don't resend to Taho when it is fallback and default is switched off
Previously toggling default to off with only Taho installed would try
to route to the previous provider, which was also Taho. This would lead
to a second popup from Taho appearing.

We now check that the new provider is not the same as the Taho provider
and, if it is, then we skip rerouting and error out.