Implement batch processing for address enrichment
Implemented a new batch processing system for address enrichment. This
aggregates enrichment lookup requests, and switches the source of data to
on-chain read only calls through multicall and a custom smart contract.
This significantly minimizes resource usage by reducing the number of
dispatched requests, improving speed and resource usage of the enrichment
process.