Remove waiting for Loading Doggo component in E2E tests
If the NFT page is loading really fast then the "loading doggo" animation
may not be visible for long enough to allow the e2e test to catch that
element. As the loader itself is not necessary for the rest of the test
let's omit testing for loader visibility, we can just test if the loader
disappeared.