Masquerade isMetaMask without breaking Taho detection
When Taho is set as default, the wallet router now reports isMetaMask =
true, while the Taho provider always reports isMetaMask = false. This
allows dApps that support Taho as a first-class provider to see it
as itself. In general the Taho provider is less concerned with
masquerading as MetaMask, leaving that to the MetaMask wrapper proxy and
the wallet router.

Additionally, the wrapper only routes an allowlist of methods to Taho to
minimize divergence of behavior from MetaMask when it is installed.