Rearrange ethereum.providers based on Taho default setting
Some frameworks, like Wagmi, always display the first item in the
providers list as the injected/browser wallet. Some sites that do this
also don't check the providers list for other wallets, so that even if
MetaMask is in the providers list, the MetaMask button shown to the user
doesn't actually bring up MetaMask, instead showing the MetaMask mobile
connection option.

To get around this, the providers list is now rearranged when Taho is
switched to or from default. When it's set as default, it is placed at
the front of the providers list; when it's set as not default, the
previous provider it replaced is set as first.