Add MetaMask mock when Taho is default and no MetaMask is present
For the specific case where Taho is set as default but MetaMask does not
seem to be installed, we now inject a MetaMask-alike in the providers
list to allow dApps that only look for MetaMask to work correctly.

When Taho is cleared from being default, the MetaMask-alike is removed
and the fact that MetaMask is not installed is left as-is for the dApp
to detect normally.