Make ci/ an independent project
This instead of a workspace in the top-level project. This allows the ci
project to have its own dependencies, rather than polluting the lockfile
for the top-level extension project with the large amount of hardhat
dependencies.

This commit basically repats what was done in
eefd47720f6156d40e53c6ee72aeb7093e400e46. Most of the changes from that commit
got overwritten during a merge and we need to bring the project config to the
desired shape.