Emit internal signers after unlocking in internal signers
This ensures that by the time internal signers hit the redux store, the
internal signers are known to be unlocked. This is useful because the
store ignores emitted signers while the signers are known to be locked,
to avoid clearing the signers list at lock time (which emits an empty
list of working signers for any services that are relying on that list
to know what can currently be actively signed).