Use JSON onboarding in e2e tests of transactions
Prior to this change we were onbording using a seed phrase, which was not ideal,
as it could lead to a seed phrase leakage in the Playwright report added to the
artifacts of CI jobs (that's why we've temporarily switched off storing of the
test artifacts).
Now that we support in Taho the onboarding with a JSON keystore file (which
stores an encrypted private key), we are good to use this method of account
import in our E2E tests. This method is safer, as the sensitive data entered in
the form during onboarding is obfuscated, so there's no risk it will leak in the
Playwright report (and even if so, what will leak will be just a password, not a
corresponding to it JSON file).
As we're switching to this onboarding method, we can bring back storing of the
Playwrigt reports in the CI jobs artifacts.

Apart from the above, as part of this commit/PR we're also introducing a script
which can be used to generate a JSON keystore file based on the provided private
key and password.