Change name of the variables storing JSON body & password
The transactions e2e tests use the first address of the testing wallet
(`testertesting.eth`). But we already know that in some tests that we plan to
add in the future, we'll want to use a different account - the third address of
the testing wallet (`e2e.testertesting.eth`). So we will need to use names of
the variables and secrets that differenciate those different addresses.