Merge `assertAssetDetailsPage()` with `verifyAssetActivityScreen()`
The `assertAssetDetailsPage` function was a more robust version of the
`verifyAssetActivityScreen` function. When `verifyAssetActivityScreen` was
created we didn't yet have the `AssetsHelper` so we kept the function checking
the activity detais in the `TransactionsHelper`. Now that we have a specific
helper for assets, we can merge both functions into one.