Remove obsolete comment
Comment suggested that we could move some of the functions from `AssetsHelper`
to `TransactionsHelper`. After thinking it through, I think it makes sense to
leave the functions where they are, especially that they're used not only during
filling of the Transaction form but also during filling of the Swap form.