Wait 1s before switching to Polygon network
We're adding a timeout, as switching to the Polygon network too fast after
importing wallet was causing some of the price info missing for the ERC-20
assets. It's quite unlikely for the real users to switch the network this fast
anyway, so the timeout does not contradict the real user behaviour.