Fix assertions made on the latest transaction
There is a stage in the tests when we want to verify the attributes of the
latest transaction. By mistake we did some of the checks on an objet that wasn't
a latest transaction, but on an object of 'any' transaction. The assertion was
failing if more than one object was detected.