Rename `verify...` functions to `assert...`
As we have functionality of verifying assets in the wallet, we recently started
to use `assert` nomenclature in the names of the new functions and methods
checking certain areas of the wallet. This was done to avoid confusion in the
naming. Now we update the older function using the `verify...` prefix, to make
the naming consistent across all our test-related files.