Selectively enable batch RPC providers
Switches to JsonRPCBatchProvider for both ankr and 1rpc. This is superb because:

- Quotas and limits are set per request as opposed to per call.
- Fixes to an extent the thundering herd issue related to enrichment which was
  causing 429s

Although ethers provides support for batch rpc providers, it does not handle
well responses where the results are not ordered. As this is in compliance
with the spec, some manual patching is necessary to process these responses
correctly.