Modify Hardhat cache config
We modify the way we manage Hardhat cache in CI.

Every time data is fetched from mainnet, Hardhat Network caches it on disk to
speed up future access. When executing on clean environments similar tests on
the Harhat network using the same forking block, we'll end up with a lot of the
same cache files produced.

The `actions/cache/save` and `actions.cache/restore` actions allow us to reuse
some files between workflow runs, allowing for faster execution. We are using
them to store and reuse the Hardhat cache, however our previous configuration
wasn't fully correct. It may be a reason why we were observing some failures of
e2e tests run on fork in CI (although we're not 100% sure yet that was the root
cause).
The `key` parameter of the `actions.cache/save` action specifies the name under
which the new cache will be saved. We don't need to include runner OS in the
name (the Hardhat cache files don't seem to be OS-specific) and we had to fix
the used file pattern, because the previously used one wasn't finding any
files, as they were present in `ci/cache`'s subdirectory
(`ci/cache/hardhat-network-fork/network-1`).
We decided also to use a forking block id in the name of the cache - we
currentny don't use other blocks than `17094516` in e2e tests, but we may use
some in the future and it may make sense to differentiate caches between them
(the Hardhat cache may differ significantly when different forking block is
used).
We also provide GH run id, to force creating new cache after each run.

We also specify the restore keys - so that when the step with the
`actions/cache/restore` action gets executed, it would pull the latest cache
matching first from the provided keys. If no matching key is found - using next
key and so on...

If the tests would be run with no cache, they would likely be failing - as it
was observed locally that the first executions of the tests always were
failing. Only once the cache was established and tests were re-run, they were
passing. That's also the reason why we moved from using `actions/cache` to
using separate `actions/cache/save` and `actions/cache/restore` actions - the
`actions/cache` didn't support saving cache after the workflow failure
(see https://github.com/actions/cache/issues/272).