Improve handling/duplication of MetaMask wrapping
For MetaMask wrappers, we now track both the unwrapped and wrapped
providers, allowing us to make sure that the overall list of providers
doesn't end up with duplicate entries for the wrapped and unwrapped
providers, or with multiple wrappers for the unwrapped provider.

Also improved is wrapping if MetaMask was injected before Taho, and
if/when MetaMask is added via the `addProvider` call. In all of these
cases, everything is deduplicated against both the wrapped and unwrapped
version of the new provider, ensuring we should only ever have one entry
in the providers list for a MetaMask-like object.