Reemit events on the Taho provider when Taho is default
Previously, when set as default, the Taho provider would capture all RPC
messages by virtue of the MetaMask wrapper proxy. However, if the dApp
attached event handlers to the injected provider before Taho was
notified that it was default (or before the user toggled it as default),
the events emitted by the Taho provider would never be seen by the dApp,
including things like chain changes and connect/disconnect events.

When Taho is the default provider, it will now reemit all events it
emits to its listeners to the last injected provider's listeners as well
(as long as there was a last injected provider). This ensures that
handlers attached before Taho took over default will still be invoked,
allowing the dApp to remain agnostic to the Taho default switch.