Fix a missed unwrap of the wrapMetaMaskProvider return
Due to how the existing window.ethereum object was being wrapped, there
was no type check to ensure that the `lastInjectedProvider` property was
in fact a provider. This was leading to a broken `lastInjectedProvider`
in cases where `window.ethereum` was defined before Taho got to it,
as the return from `wrapMetaMaskProvider` has two properties, only one
of which is the provider to be used.

Fix this by pulling out `wrappedLastInjectedProvider` to handle type
checking, and unwrapping the `wrapMetaMaskProvider` return value.