Refactor function closing Accounts slide up
As after opening the Accounts list, there were multiple `slide_up_menu`
elements in the wallet's DOM, and only one was not of a `closed` class, we were
using the `ElementHandle` syntax to find the one which is not closed (the one
we wanted to click). This then influenced the way how we can locate the `Close
menu` button, making the code quite complicated.
In this PR we are simplifying the code, by filtering the slide up menus by the
`Accounts` text. Then we are able to point to the right `Close menu` element by
using `.first()` locator.