Move the `closeAccountsSlideUp()` steps under `addAddressToAccount()`
Both functions are currently not used by the e2e tests, but we want to keep
them in case they're needed in the future. We're moving the steps closing the
Accounts slide up to the `addAddressToAccount()` functions for now. They may be
separated on the future, if there will be a need for this.