Add _metamask to mocked MetaMask properties when mocking
This continues to be managed in the window provider's mocking paths,
but the cases where this is pulled from the metaMaskMock are expanded.
When Taho is set as default, MetaMask-specific properties are pulled
from the mock. This was already effectively happening for `isMetaMask`,
but is now expanded to include _metamask and _state, which are also
sometimes used to recognize MetaMask.