Add close callback to `showExtensionPopup` helper
The `showExtensionPopup` helper previously returned the created window
object, but offered no easy way to monitor for the closing of the
window. We monitor for such a closing in more global ways for signature
requests, mediated by the Main service; however, there are use cases
where we need to manage internal state in the ProviderBridge service in
a self-contained way.

To allow this, a new `onClose` parameter is added to
`showExtensionPopup`; it is a function that will be called when the
popup window created by the function call is closed, no matter how it
was closed.