Clear pending permission request on permission popover close
This effectively rejects the permission request, and allows for future
permission requests to come through normally. Previously, because this
state was not cleared, the request was not treated as rejected from the
dApp perspective (meaning the dApp would stay in a pending state that
required a reload to recover from), and due to recent changes to
disallow multiple parallel permission requests, future requests were
immediately denied.