Group private keys when categorizing
When the signer id abstraction was introduced, I misunderstood each
private key as needing to be treated as its own signer. In fact, we
treat, on the UI side, all private keys as a single signer group. To
fix this, rather than return the wallet id for the signer in these
cases, we just return the constant string `"private-key"`. This ensures
all private keys are categorized together.