Track data signature events and failures when analytics are enabled
Nothing is recorded about the signature other than the fact that data
was signed. Failures track error reason, which is a very limited set of
possible errors managed by the `SigningService`.

Notably, this commit also moves handling for transaction signatures out
of `Main` and into `AnalyticsService`, which should be loosely coupled
to service events when possible.