Add support for Mainnet Sepolia testnet
Although it was previously possible to add Sepolia testnet as a custom network,
we want to add it to the list of networks automatically added to the networks
list upon switching on the 'Show testnet networks' setting. We want that because
the only testnet currently added to that list is Goerli, which will become
deprecated at the end of 2023 (see
https://ethereum-magicians.org/t/proposal-predictable-ethereum-testnet-lifecycle/11575).
So far we want to still keep Goerli on the list, as it is still widely used.
We'll need to remove it once it'll reach end of its life.

As at the moment `0x` project which we use for swaps does not support Sepolia,
the sweep functionality will be disabled on this testnet. We need to enable it
once the support is added.