Add a skeleton for the portal Banner
We use a dismissable item for the banner itself, which will be dismissed
on clicking the Explore The Island link.

No styling is included at the moment, and there is a stub for not
displaying the banner unless the token has been detected as
deployed---this part is not yet implemented.