Update e2e tests after change of testing accounts
We've rotated the wallets used for testing of the extension and moved some of
the assets. The GH secrets got already updated accordingly, we now need to
update the addresses of wallets and change some other values used in tests, to
fit the assets that we have on the new testing wallets.
Changes include:
* switching of `e2e.testertesting.eth` wallet from
  `0x6f1b1f1feb01235e15a7962f16c389c7f8218ed6` to
  `0x9d373acbe8540895fa1752ab463ab31bbab2b38f`
* switching of `testertesting.eth` wallet from
  `0x0581470a8b62bd35dbf121a6329d43e7edd20fc7` to
  `0x6e80164ea60673d64d5d6228beb684a1274bb017`
* changing the POAP used in the NFTs tests
* changing the number of a forking block (the new testing wallet didn't have any
  assets yet on an old forking block) and switching to a different asset used in
  tests of ERC-20 transfers.