Add vestigial wallet_requestPermissions support
wallet_requestPermissions is defined by EIP-2255, but does not seem to
have much use outside of an alternative to `eth_accounts` and a way to
re-prompt the user for access to more/different accounts.

For the time being, `wallet_requestPermissions` is implemented exactly
like `eth_requestAccounts`, and `wallet_getPermissions` exactly like
`eth_accounts`, except that the return value is structured somewhat
differently.

If and when Taho supports connecting multiple accounts to the same dApp
simultaneously, we can extend this model with two things:

- Always prompting the user to update their connected accounts when
  `wallet_requestPermissions` is invoked.
- Returning more than one account in the `wallet_getPermissions`
  `restrictedReturnedAccounts` caveat.