Modify locator identifying `Select token` buttons
It looks that after upgrading to the latest version of Playwright something
has changed in the way how some locators work. The `.getByRole("button", { name:
"Select token", exact: true })` locator no longer identified the button we
wanted to click on the Swap screen. In this commit we're changing the location
method.