Expect the SIWE test to fail, due to bug on login.xyz
Recently the signing in with Ethereum on login.xyz started to misbehave. The old
behavior was that there was a `Vote for your favorite emoji` poll visible on the
website after confirming SIWE in the wallet. Now the website shows the same
state as before the signing. In website console there's a `Cannot read
properties of undefined (reading 'vote')` error. The issue got reported on
https://discord.com/channels/862419652286218251/886997073650655232/1173226370776694794.
We've been using login.xyz in our automated tests when checking SIWE. Until the
issue is fixed, we're marking the test as expected to fail. That means that the
test will be marked in logs as passed when there's no `Vote for your favorite
emoji` on the page after signing. If the string will be found, Playwright will
log a failure and will print `Expected to fail, but passed.`. Placing the
`test.fail()` expression just before the failing step makes sure that failures
on earlier stages will not be considered as expected and will still be
flagged.