Use soft assertions when expecting snackbar
There is a bug in the wallet which manifests with the snackbar with the action's
status not displaying after that action. This happens sporadically and does not
mean that the action wasn't successful. We want to be notified when the bug
occurs, but we don't want it to block execution of the subsequent steps. That's
why we use the soft assertions.