Remove Goerli testnet from the testnets list
Each Ethereum testnet has a limited lifespan. With end of 2023 Goerli testnet
becomes deprecated. As it's one of the testnets we have set as a default
testnet, we will no longer want to show it after that date.
The remaining default testnets will be Ethereum Sepolia and Arbitrum Sepolia.
Other tesnets can be added by users manually (no changes in that area).