Move action using node12 to a newer version
Node12 has been out of support since April 2022. As a result GH has started the
deprecation process of Node12 for GitHub Actions by migrating all actions to run
on Node16 ([read
more](https://github.blog/changelog/2023-07-17-github-actions-removal-of-node12-from-the-actions-runner/)).
The `softprops/action-gh-release` action's version which we used in the
`build.yml` workflow was running on Node12. But a newer version (`v0.1.15`) of
the action is run on Node16. We're swithing to using that newer version. This
should make `The following actions uses node12 which is deprecated and will be
forced to run on node16: softprops/action-gh-release@1e07f4...` annotation
disappear from the details of the workflow runs.

The upgrade may also remove the `The "set-output" command is deprecated and will
be disabled soon. Please upgrade to using Environment Files. For more
information see:
https://github.blog/changelog/2022-10-11-github-actions-deprecating-save-state-and-set-output-commands/`
annotation.